3.0
This is a much more robust version that has more randomness, and movement, that are dynamic features, we added more balloons and updated our balloon move function and fixed our clustering to be more dynamic as well. These tests that also have more balloons and randomness tests show that the number of arrows can sometimes be a little as 4, or as much as 9 depending on placement, and maybe higher or lower if the random generator places them in such a manor.